refeat: main_window repartido em várias funções separdas para ficar mais clara
